+++ Fixtures/iOS/AppWithExtensions/WatchApp Extension/HostingController.swift
@@ -1,0 +1,9 @@
+import WatchKit
+import Foundation
+import SwiftUI
+
+class HostingController: WKHostingController<ContentView> {
+    override var body: ContentView {
+        return ContentView()
+    }
+}