+++ Sources/XcodeProj/Scheme/XCScheme+Runnable.swift
@@ -32,10 +32,13 @@ extension XCScheme {
         }
 
         // MARK: - Equatable
+        func isEqual(other: Runnable) -> Bool {
+            return runnableDebuggingMode == other.runnableDebuggingMode &&
+                buildableReference == other.buildableReference
+        }
 
         public static func == (lhs: Runnable, rhs: Runnable) -> Bool {
-            return lhs.runnableDebuggingMode == rhs.runnableDebuggingMode &&
-                lhs.buildableReference == rhs.buildableReference
+            return lhs.isEqual(other: rhs) && rhs.isEqual(other: lhs)
         }
     }
 }