+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - Support `on-demand-install-capable` application https://github.com/tuist/XcodeProj/pull/554 by @d-date
+- Add RemotePath to RemoteRunnable https://github.com/tuist/XcodeProj/pull/555 by @kwridan
 
 ## 7.12.0